+++ packages/directory/addon/components/dir-item-favorite-cell.hbs
@@ -1,0 +1,16 @@
+{{!-- Copyright 2020, Yahoo Holdings Inc. Licensed under the terms of the MIT license. See accompanying LICENSE.md file for terms. --}}
+{{#if @value.isFavorite}}
+  <DenaliIcon
+    class="link p-x-30"
+    @icon="star-solid"
+    @size="small"
+    {{on "click" (fn this.toggleFavorite @value.isFavorite)}}
+  />
+{{else}}
+  <DenaliIcon
+    class="link p-x-30"
+    @icon="star"
+    @size="small"
+    {{on "click" (fn this.toggleFavorite @value.isFavorite)}}
+  />
+{{/if}}
