+++ packages/directory/app/components/dir-item-favorite-cell.js
@@ -1,0 +1,1 @@
+export { default } from 'navi-directory/components/dir-item-favorite-cell';
